+++ src/pages/authentication/forgotPassword/validate.js
@@ -1,0 +1,22 @@
+import { regex } from '../../../utils';
+const validate = values => {
+  const errors = {};
+  const requiredFields = [
+    'firstName',
+    'lastName',
+    'email',
+    'password',
+    'confirm-password',
+  ];
+  requiredFields.forEach(field => {
+    if (!values[field]) {
+      errors[field] = 'This field is required';
+    }
+  });
+  if (values.email && !regex.EMAIL_REGEX.test(values.email)) {
+    errors.email = 'Invalid email address';
+  }
+  return errors;
+};
+
+export default validate;
