+++ src/pages/authentication/login/LoginForm.js
@@ -1,45 +1,9 @@
 import React from 'react';
 import { Field, reduxForm } from 'redux-form';
-import TextField from 'material-ui/TextField';
 import RaisedButton from 'material-ui/RaisedButton';
 import './formStyle.css';
-import { regex } from '../../../utils';
-
-const validate = values => {
-  const errors = {};
-  const requiredFields = ['email', 'password'];
-  requiredFields.forEach(field => {
-    if (!values[field]) {
-      errors[field] = 'This field is required';
-    }
-  });
-  if (values.email && !regex.EMAIL_REGEX.test(values.email)) {
-    errors.email = 'Invalid email address';
-  }
-  if (values.password && !regex.passwordRegex.test(values.password)) {
-    errors.password =
-      'Password must contains at least 6 character include number and special character ';
-  }
-  return errors;
-};
-
-const renderField = ({
-  input,
-  className,
-  label,
-  type,
-  meta: { touched, error },
-  ...custom
-}) => (
-  <TextField
-    floatingLabelText={label}
-    errorText={touched && error}
-    type={type}
-    {...input}
-    {...custom}
-    className={className}
-  />
-);
+import { renderField } from '../../../utils';
+import validate from './validate';
 
 const LoginForm = ({
   error,
