+++ src/pages/authentication/register/RegisterForm.js
@@ -1,58 +1,9 @@
 import React from 'react';
 import { Field, reduxForm } from 'redux-form';
-import TextField from 'material-ui/TextField';
 import RaisedButton from 'material-ui/RaisedButton';
-import { regex } from '../../../utils';
 import './RegisterForm.css';
-
-const validate = values => {
-  const errors = {};
-  const requiredFields = [
-    'firstName',
-    'lastName',
-    'email',
-    'password',
-    'confirmPassword',
-  ];
-  requiredFields.forEach(field => {
-    if (!values[field]) {
-      errors[field] = 'This field is required!';
-    }
-  });
-  if (values.email && !regex.EMAIL_REGEX.test(values.email)) {
-    errors.email = 'Invalid email address';
-  }
-  if (
-    values.password &&
-    values.confirmPassword &&
-    values.password !== values.confirmPassword
-  ) {
-    errors.confirmPassword = 'Password does not match!';
-  }
-  if (values.password && !regex.passwordRegex.test(values.password)) {
-    errors.password =
-      'Password must contains at least 6 character include number and special character ';
-  }
-  return errors;
-};
-
-const renderField = ({
-  input,
-  label,
-  type,
-  className,
-  meta: { touched, error },
-  ...custom
-}) => (
-  <TextField
-    floatingLabelText={label}
-    errorText={touched && error}
-    type={type}
-    {...input}
-    {...custom}
-    className={className}
-  />
-);
+import validate from './validate';
+import { renderField } from '../../../utils';
 
 const RegisterForm = ({
   error,
