+++ src/utils/index.js
@@ -1,1 +1,2 @@
 export { default as regex } from './regex';
+export { default as renderField } from './field';
